build requests
